[FIX] Moved the wrong part of the Windows service.
